Ignore broken lines.
Currently, contributions-graphs will throw/error out if it finds
and input that has a poorly formated line.  For the uses I care
about, I would rather just have this code warn and continue.

If I plan to send a proper CL, we'd need to fix the test cases
at a minimum.